update surface control
